application/python/requirements.txt
flask==3.0.0
psycopg2-binary==2.9.9
redis==5.0.1
prometheus-client==0.19.0
python-json-logger==2.0.7
requests==2.31.0
pytest==7.4.3
pytest-cov==4.1.0
